Add notion of an "enemy tribe" to DB, add endpoints for managing enemy tribes, added option to force credentials for testing, finish first pass at user stats endpoint, move "nuke", "fang", "fake", etc. classification/values to ArmyStats utilities, start adding "backline DVs" to admin tribe summary

+++ TW.Vault/Model/JSON/UserStats.cs
@@ -12,6 +12,6 @@ namespace TW.Vault.Model.JSON
         public int NukesInPastWeek { get; set; }
         public int FangsInPastWeek { get; set; }
         public int FakesInPastWeek { get; set; }
-        public Dictionary<String, long> PopPerTribe { get; set; }
+        public Dictionary<String, float> PopPerTribe { get; set; }
     }
 }
